refactor(activities): split lib/events.ts into lib/activities/ folder
- lib/activities/types.ts  — AppEvent interface
- lib/activities/utils.ts  — NEW_ACTIVITY_CUTOFF_MS, toEventIso, isNewActivity
- lib/activities/hooks.ts  — useEvents hook
- lib/activities/index.ts  — re-exports all public API
- Delete lib/events.ts, update all imports
