+++ frontend/src/lib/activities/index.ts
@@ -1,0 +1,3 @@
+export type { AppEvent } from "./types";
+export { NEW_ACTIVITY_CUTOFF_MS, isNewActivity, toEventIso } from "./utils";
+export { useEvents } from "./hooks";
