+++ frontend/src/lib/activities/types.ts
@@ -1,0 +1,8 @@
+export interface AppEvent {
+  id: number;
+  ts: string;
+  kind: string;
+  actor: string | null;
+  target: string | null;
+  payload: Record<string, unknown>;
+}
